Merging mooo's recent changes

+++ tests/hash-target.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2014-2015, The Monero Project
+// Copyright (c) 2014-2018, The Monero Project
 // 
 // All rights reserved.
 // 
@@ -33,7 +33,7 @@
 #include <cstring>
 #include <limits>
 #include "crypto/hash.h"
-#include "cryptonote_core/difficulty.h"
+#include "cryptonote_basic/difficulty.h"
 
 using namespace std;
 using cryptonote::check_hash;
@@ -41,7 +41,7 @@ using cryptonote::check_hash;
 int main(int argc, char *argv[]) {
   crypto::hash h;
   for (uint64_t diff = 1;; diff += 1 + (diff >> 8)) {
-    for (int b = 0; b < 256; b++) {
+    for (uint16_t b = 0; b < 256; b++) {
       memset(&h, b, sizeof(crypto::hash));
       if (check_hash(h, diff) != (b == 0 || diff <= 255 / b)) {
         return 1;
